Squashed 'vision' changes from 744f8cf5b..432738226 (#1157)
432738226 Merge pull request #801 from liyuanz/fix_std
1f88d5340 update
6c90a42dd update
7f61ee220 update
9d0f11fff update
7cb776560 Merge pull request #800 from sushmita/sbm_2025_fix
609d49549 added stacksize and heapsize
ee0644598 Added stacksize and heapsize
1ad4a7508  2025_tool_ver_hang_fix
1004d651f Merge pull request #796 from liyuanz/remove_sw_emu_next
d130217ca update
a9315fdd1 update
b08990bb4 update Makefile
86e7e073a remove sw_emu in description.json
1ff3bbcc8 Merge pull request #795 from liyuanz/change_python3
64b7025b7 change Makefile and utils.mk
72190a7f6 change 2024.2_stable_latest to 2025.1_stable_latest

+++ vision/L1/include/aie-ml/imgproc/xf_sbm_impl.hpp
@@ -308,7 +308,7 @@ SbmBaseImpl<TILE_OUT_HEIGHT,
 
     acc2 = ::aie::mul(first_min, hundred);
 
-    acc2 = ::aie::mac(acc2, first_min, UNIQUENESS_RATIO);
+    acc2 = ::aie::mac(acc2, first_min, (uint16_t)UNIQUENESS_RATIO);
 
     auto bcond = ::aie::le(acc1.template to_vector<int32_t>(0), acc2.template to_vector<int32_t>(0));
